Add blocksapi blocks loader
